Update the pubspec.yaml file
This commit updates the settings in the dependency file of this package.
First, it sets the commit hash for the plugin version made available by
Github. This change grants that a new committed version will not break
the package.

In addition, it changes the Dart version to a more widely used version
and removes a patch version of the HTTP package to be compatible with
the new Dart version.

See-more: https://dart.dev/tools/pub/dependencies